+++ controllers/redis_controller.go
@@ -27,7 +27,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	redisv1beta1 "redis-operator/api/v1beta1"
 )
@@ -52,7 +51,7 @@ func (r *RedisReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 		}
 		return ctrl.Result{}, err
 	}
-	
+
 	if err := k8sutils.HandleRedisFinalizer(instance, r.Client); err != nil {
 		return ctrl.Result{}, err
 	}
@@ -61,14 +60,11 @@ func (r *RedisReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 		return ctrl.Result{}, err
 	}
 
-	if err := controllerutil.SetControllerReference(instance, instance, r.Scheme); err != nil {
-		return ctrl.Result{}, err
-	}
-	err = k8sutils.CreateStandAloneRedis(instance)
+	err = k8sutils.CreateStandaloneRedis(instance)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
-	err = k8sutils.CreateStandAloneService(instance)
+	err = k8sutils.CreateStandaloneService(instance)
 	if err != nil {
 		return ctrl.Result{}, err
 	}