+++ k8sutils/redis-cluster.go
@@ -1,8 +1,9 @@
 package k8sutils
 
 import (
-	corev1 "k8s.io/api/core/v1"
 	redisv1beta1 "redis-operator/api/v1beta1"
+
+	corev1 "k8s.io/api/core/v1"
 )
 
 // RedisClusterSTS is a interface to call Redis Statefulset function
@@ -10,6 +11,8 @@ type RedisClusterSTS struct {
 	RedisStateFulType string
 	ExternalConfig    *string
 	Affinity          *corev1.Affinity `json:"affinity,omitempty"`
+	ReadinessProbe    *redisv1beta1.Probe
+	LivenessProbe     *redisv1beta1.Probe
 }
 
 // RedisClusterService is a interface to call Redis Service function
@@ -18,9 +21,10 @@ type RedisClusterService struct {
 }
 
 // generateRedisStandalone generates Redis standalone information
-func generateRedisClusterParams(cr *redisv1beta1.RedisCluster, replicas *int32, externalConfig *string, affinity *corev1.Affinity) statefulSetParameters {
+func generateRedisClusterParams(cr *redisv1beta1.RedisCluster, replicas int32, externalConfig *string, affinity *corev1.Affinity) statefulSetParameters {
 	res := statefulSetParameters{
-		Replicas:          replicas,
+		Metadata:          cr.ObjectMeta,
+		Replicas:          &replicas,
 		NodeSelector:      cr.Spec.NodeSelector,
 		SecurityContext:   cr.Spec.SecurityContext,
 		PriorityClassName: cr.Spec.PriorityClassName,
@@ -43,7 +47,7 @@ func generateRedisClusterParams(cr *redisv1beta1.RedisCluster, replicas *int32, 
 }
 
 // generateRedisStandaloneContainerParams generates Redis container information
-func generateRedisClusterContainerParams(cr *redisv1beta1.RedisCluster) containerParameters {
+func generateRedisClusterContainerParams(cr *redisv1beta1.RedisCluster, readinessProbeDef *redisv1beta1.Probe, livenessProbeDef *redisv1beta1.Probe) containerParameters {
 	trueProperty := true
 	falseProperty := false
 	containerProp := containerParameters{
@@ -72,9 +76,19 @@ func generateRedisClusterContainerParams(cr *redisv1beta1.RedisCluster) containe
 		}
 
 	}
+	if readinessProbeDef != nil {
+		containerProp.ReadinessProbe = readinessProbeDef
+	}
+	if livenessProbeDef != nil {
+		containerProp.LivenessProbe = livenessProbeDef
+	}
 	if cr.Spec.Storage != nil {
 		containerProp.PersistenceEnabled = &trueProperty
 	}
+	if cr.Spec.TLS != nil {
+		containerProp.TLSConfig = cr.Spec.TLS
+	}
+
 	return containerProp
 }
 
@@ -83,6 +97,8 @@ func CreateRedisLeader(cr *redisv1beta1.RedisCluster) error {
 	prop := RedisClusterSTS{
 		RedisStateFulType: "leader",
 		Affinity:          cr.Spec.RedisLeader.Affinity,
+		ReadinessProbe:    cr.Spec.RedisLeader.ReadinessProbe,
+		LivenessProbe:     cr.Spec.RedisLeader.LivenessProbe,
 	}
 	if cr.Spec.RedisLeader.RedisConfig != nil {
 		prop.ExternalConfig = cr.Spec.RedisLeader.RedisConfig.AdditionalRedisConfig
@@ -95,6 +111,8 @@ func CreateRedisFollower(cr *redisv1beta1.RedisCluster) error {
 	prop := RedisClusterSTS{
 		RedisStateFulType: "follower",
 		Affinity:          cr.Spec.RedisFollower.Affinity,
+		ReadinessProbe:    cr.Spec.RedisFollower.ReadinessProbe,
+		LivenessProbe:     cr.Spec.RedisFollower.LivenessProbe,
 	}
 	if cr.Spec.RedisFollower.RedisConfig != nil {
 		prop.ExternalConfig = cr.Spec.RedisFollower.RedisConfig.AdditionalRedisConfig
@@ -118,34 +136,24 @@ func CreateRedisFollowerService(cr *redisv1beta1.RedisCluster) error {
 	return prop.CreateRedisClusterService(cr)
 }
 
-func (service RedisClusterSTS) getReplicaCount(cr *redisv1beta1.RedisCluster) *int32 {
-	var replicas *int32
-	if service.RedisStateFulType == "leader" {
-		replicas = cr.Spec.RedisLeader.Replicas
-	} else {
-		replicas = cr.Spec.RedisFollower.Replicas
-	}
-
-	// We fall back to the overall/default size if we don't have a specific one.
-	if replicas == nil {
-		replicas = cr.Spec.Size
-	}
-	return replicas
+func (service RedisClusterSTS) getReplicaCount(cr *redisv1beta1.RedisCluster) int32 {
+	return cr.Spec.GetReplicaCounts(service.RedisStateFulType)
 }
 
 // CreateRedisClusterSetup will create Redis Setup for leader and follower
 func (service RedisClusterSTS) CreateRedisClusterSetup(cr *redisv1beta1.RedisCluster) error {
 	stateFulName := cr.ObjectMeta.Name + "-" + service.RedisStateFulType
-	logger := stateFulSetLogger(cr.Namespace, stateFulName)
-	labels := getRedisLabels(stateFulName, "cluster", service.RedisStateFulType)
-	objectMetaInfo := generateObjectMetaInformation(stateFulName, cr.Namespace, labels, generateStatefulSetsAnots())
+	logger := statefulSetLogger(cr.Namespace, stateFulName)
+	labels := getRedisLabels(stateFulName, "cluster", service.RedisStateFulType, cr.ObjectMeta.Labels)
+	annotations := generateStatefulSetsAnots(cr.ObjectMeta)
+	objectMetaInfo := generateObjectMetaInformation(stateFulName, cr.Namespace, labels, annotations)
 	err := CreateOrUpdateStateFul(
 		cr.Namespace,
 		objectMetaInfo,
-		labels,
 		generateRedisClusterParams(cr, service.getReplicaCount(cr), service.ExternalConfig, service.Affinity),
 		redisClusterAsOwner(cr),
-		generateRedisClusterContainerParams(cr),
+		generateRedisClusterContainerParams(cr, service.ReadinessProbe, service.LivenessProbe),
+		cr.Spec.Sidecars,
 	)
 	if err != nil {
 		logger.Error(err, "Cannot create statefulset for Redis", "Setup.Type", service.RedisStateFulType)
@@ -158,29 +166,22 @@ func (service RedisClusterSTS) CreateRedisClusterSetup(cr *redisv1beta1.RedisClu
 func (service RedisClusterService) CreateRedisClusterService(cr *redisv1beta1.RedisCluster) error {
 	serviceName := cr.ObjectMeta.Name + "-" + service.RedisServiceRole
 	logger := serviceLogger(cr.Namespace, serviceName)
-	labels := getRedisLabels(serviceName, "cluster", service.RedisServiceRole)
+	labels := getRedisLabels(serviceName, "cluster", service.RedisServiceRole, cr.ObjectMeta.Labels)
+	annotations := generateServiceAnots(cr.ObjectMeta)
 	if cr.Spec.RedisExporter != nil && cr.Spec.RedisExporter.Enabled {
 		enableMetrics = true
 	}
-	objectMetaInfo := generateObjectMetaInformation(serviceName, cr.Namespace, labels, generateServiceAnots())
-	headlessObjectMetaInfo := generateObjectMetaInformation(serviceName+"-headless", cr.Namespace, labels, generateServiceAnots())
-	err := CreateOrUpdateHeadlessService(cr.Namespace, headlessObjectMetaInfo, labels, redisClusterAsOwner(cr))
+	objectMetaInfo := generateObjectMetaInformation(serviceName, cr.Namespace, labels, annotations)
+	headlessObjectMetaInfo := generateObjectMetaInformation(serviceName+"-headless", cr.Namespace, labels, annotations)
+	err := CreateOrUpdateService(cr.Namespace, headlessObjectMetaInfo, redisClusterAsOwner(cr), false, true)
 	if err != nil {
 		logger.Error(err, "Cannot create headless service for Redis", "Setup.Type", service.RedisServiceRole)
 		return err
 	}
-	err = CreateOrUpdateService(cr.Namespace, objectMetaInfo, labels, redisClusterAsOwner(cr), enableMetrics)
+	err = CreateOrUpdateService(cr.Namespace, objectMetaInfo, redisClusterAsOwner(cr), enableMetrics, false)
 	if err != nil {
 		logger.Error(err, "Cannot create service for Redis", "Setup.Type", service.RedisServiceRole)
 		return err
 	}
 	return nil
-}
-
-func getRedisLabels(name, setupType, role string) map[string]string {
-	return map[string]string{
-		"app":              name,
-		"redis_setup_type": setupType,
-		"role":             role,
-	}
 }