+++ k8sutils/statefulset.go
@@ -2,24 +2,28 @@ package k8sutils
 
 import (
 	"context"
+	"fmt"
+	"path"
+	redisv1beta1 "redis-operator/api/v1beta1"
 	"sort"
 
 	"github.com/banzaicloud/k8s-objectmatcher/patch"
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
 	redisExporterContainer = "redis-exporter"
-	graceTime              = 15
 )
 
 // statefulSetParameters will define statefulsets input params
 type statefulSetParameters struct {
 	Replicas              *int32
+	Metadata              metav1.ObjectMeta
 	NodeSelector          map[string]string
 	SecurityContext       *corev1.PodSecurityContext
 	PriorityClassName     string
@@ -45,38 +49,57 @@ type containerParameters struct {
 	SecretName                   *string
 	SecretKey                    *string
 	PersistenceEnabled           *bool
+	TLSConfig                    *redisv1beta1.TLSConfig
+	ReadinessProbe               *redisv1beta1.Probe
+	LivenessProbe                *redisv1beta1.Probe
 }
 
 // CreateOrUpdateStateFul method will create or update Redis service
-func CreateOrUpdateStateFul(namespace string, stsMeta metav1.ObjectMeta, labels map[string]string, params statefulSetParameters, ownerDef metav1.OwnerReference, containerParams containerParameters) error {
-	logger := stateFulSetLogger(namespace, stsMeta.Name)
-	storedStateful, err := GetStateFulSet(namespace, stsMeta.Name)
-	statefulSetDef := generateStateFulSetsDef(stsMeta, labels, params, ownerDef, containerParams)
+func CreateOrUpdateStateFul(namespace string, stsMeta metav1.ObjectMeta, params statefulSetParameters, ownerDef metav1.OwnerReference, containerParams containerParameters, sidecars *[]redisv1beta1.Sidecar) error {
+	logger := statefulSetLogger(namespace, stsMeta.Name)
+	storedStateful, err := GetStatefulSet(namespace, stsMeta.Name)
+	statefulSetDef := generateStatefulSetsDef(stsMeta, params, ownerDef, containerParams, getSidecars(sidecars))
 	if err != nil {
 		if err := patch.DefaultAnnotator.SetLastAppliedAnnotation(statefulSetDef); err != nil {
 			logger.Error(err, "Unable to patch redis statefulset with comparison object")
 			return err
 		}
 		if errors.IsNotFound(err) {
-			return createStateFulSet(namespace, statefulSetDef)
+			return createStatefulSet(namespace, statefulSetDef)
 		}
 		return err
 	}
-	return patchStateFulSet(storedStateful, statefulSetDef, namespace)
+	return patchStatefulSet(storedStateful, statefulSetDef, namespace)
 }
 
 // patchStateFulSet will patch Redis Kubernetes StateFulSet
-func patchStateFulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.StatefulSet, namespace string) error {
-	logger := stateFulSetLogger(namespace, storedStateful.Name)
-	patchResult, err := patch.DefaultPatchMaker.Calculate(storedStateful, newStateful)
+func patchStatefulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.StatefulSet, namespace string) error {
+	logger := statefulSetLogger(namespace, storedStateful.Name)
+
+	// We want to try and keep this atomic as possible.
+	newStateful.ResourceVersion = storedStateful.ResourceVersion
+	newStateful.CreationTimestamp = storedStateful.CreationTimestamp
+	newStateful.ManagedFields = storedStateful.ManagedFields
+
+	patchResult, err := patch.DefaultPatchMaker.Calculate(storedStateful, newStateful,
+		patch.IgnoreStatusFields(),
+		patch.IgnoreVolumeClaimTemplateTypeMetaAndStatus(),
+		patch.IgnoreField("kind"),
+		patch.IgnoreField("apiVersion"),
+	)
 	if err != nil {
 		logger.Error(err, "Unable to patch redis statefulset with comparison object")
 		return err
 	}
 	if !patchResult.IsEmpty() {
-		newStateful.ResourceVersion = storedStateful.ResourceVersion
-		newStateful.CreationTimestamp = storedStateful.CreationTimestamp
-		newStateful.ManagedFields = storedStateful.ManagedFields
+		logger.Info("Changes in statefulset Detected, Updating...", "patch", string(patchResult.Patch))
+		// Field is immutable therefore we MUST keep it as is.
+		if !apiequality.Semantic.DeepEqual(newStateful.Spec.VolumeClaimTemplates, storedStateful.Spec.VolumeClaimTemplates) {
+			logger.Error(fmt.Errorf("ignored change in cr.spec.storage.volumeClaimTemplate because it is not supported by statefulset"),
+				"Redis statefulset is patched partially")
+			newStateful.Spec.VolumeClaimTemplates = storedStateful.Spec.VolumeClaimTemplates
+		}
+
 		for key, value := range storedStateful.Annotations {
 			if _, present := newStateful.Annotations[key]; !present {
 				newStateful.Annotations[key] = value
@@ -86,26 +109,29 @@ func patchStateFulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.St
 			logger.Error(err, "Unable to patch redis statefulset with comparison object")
 			return err
 		}
-		return updateStateFulSet(namespace, newStateful)
+		return updateStatefulSet(namespace, newStateful)
 	}
+	logger.Info("Reconciliation Complete, no Changes required.")
 	return nil
 }
 
-// generateStateFulSetsDef generates the statefulsets definition of Redis
-func generateStateFulSetsDef(stsMeta metav1.ObjectMeta, labels map[string]string, params statefulSetParameters, ownerDef metav1.OwnerReference, containerParams containerParameters) *appsv1.StatefulSet {
+// generateStatefulSetsDef generates the statefulsets definition of Redis
+func generateStatefulSetsDef(stsMeta metav1.ObjectMeta, params statefulSetParameters, ownerDef metav1.OwnerReference, containerParams containerParameters, sidecars []redisv1beta1.Sidecar) *appsv1.StatefulSet {
 	statefulset := &appsv1.StatefulSet{
 		TypeMeta:   generateMetaInformation("StatefulSet", "apps/v1"),
 		ObjectMeta: stsMeta,
 		Spec: appsv1.StatefulSetSpec{
-			Selector:    LabelSelectors(labels),
+			Selector:    LabelSelectors(stsMeta.GetLabels()),
 			ServiceName: stsMeta.Name,
 			Replicas:    params.Replicas,
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
+					Labels:      stsMeta.GetLabels(),
+					Annotations: generateStatefulSetsAnots(stsMeta),
+					// Annotations: stsMeta.Annotations,
 				},
 				Spec: corev1.PodSpec{
-					Containers:        generateContainerDef(stsMeta.Name, containerParams, params.EnableMetrics, params.ExternalConfig),
+					Containers:        generateContainerDef(stsMeta.GetName(), containerParams, params.EnableMetrics, params.ExternalConfig, sidecars),
 					NodeSelector:      params.NodeSelector,
 					SecurityContext:   params.SecurityContext,
 					PriorityClassName: params.PriorityClassName,
@@ -121,11 +147,22 @@ func generateStateFulSetsDef(stsMeta metav1.ObjectMeta, labels map[string]string
 		statefulset.Spec.Template.Spec.ImagePullSecrets = *params.ImagePullSecrets
 	}
 	if containerParams.PersistenceEnabled != nil && *containerParams.PersistenceEnabled {
-		statefulset.Spec.VolumeClaimTemplates = append(statefulset.Spec.VolumeClaimTemplates, createPVCTemplate(stsMeta.Name, params.PersistentVolumeClaim))
+		statefulset.Spec.VolumeClaimTemplates = append(statefulset.Spec.VolumeClaimTemplates, createPVCTemplate(stsMeta, params.PersistentVolumeClaim))
 	}
 	if params.ExternalConfig != nil {
 		statefulset.Spec.Template.Spec.Volumes = getExternalConfig(*params.ExternalConfig)
 	}
+
+	if containerParams.TLSConfig != nil {
+		statefulset.Spec.Template.Spec.Volumes = append(statefulset.Spec.Template.Spec.Volumes,
+			corev1.Volume{
+				Name: "tls-certs",
+				VolumeSource: corev1.VolumeSource{
+					Secret: &containerParams.TLSConfig.Secret,
+				},
+			})
+	}
+
 	AddOwnerRefToObject(statefulset, ownerDef)
 	return statefulset
 }
@@ -147,41 +184,110 @@ func getExternalConfig(configMapName string) []corev1.Volume {
 }
 
 // createPVCTemplate will create the persistent volume claim template
-func createPVCTemplate(name string, storageSpec corev1.PersistentVolumeClaim) corev1.PersistentVolumeClaim {
+func createPVCTemplate(stsMeta metav1.ObjectMeta, storageSpec corev1.PersistentVolumeClaim) corev1.PersistentVolumeClaim {
 	pvcTemplate := storageSpec
 	pvcTemplate.CreationTimestamp = metav1.Time{}
-	pvcTemplate.Name = name
+	pvcTemplate.Name = stsMeta.GetName()
+	pvcTemplate.Labels = stsMeta.GetLabels()
+	// We want the same annoations as the StatefulSet here
+	pvcTemplate.Annotations = generateStatefulSetsAnots(stsMeta)
 	if storageSpec.Spec.AccessModes == nil {
 		pvcTemplate.Spec.AccessModes = []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce}
 	} else {
 		pvcTemplate.Spec.AccessModes = storageSpec.Spec.AccessModes
 	}
+	pvcVolumeMode := corev1.PersistentVolumeFilesystem
+	if storageSpec.Spec.VolumeMode != nil {
+		pvcVolumeMode = *storageSpec.Spec.VolumeMode
+	}
+	pvcTemplate.Spec.VolumeMode = &pvcVolumeMode
 	pvcTemplate.Spec.Resources = storageSpec.Spec.Resources
-	pvcTemplate.Spec.Selector = storageSpec.Spec.Selector
 	pvcTemplate.Spec.Selector = storageSpec.Spec.Selector
 	return pvcTemplate
 }
 
-// generateContainerDef generates container fefinition for Redis
-func generateContainerDef(name string, containerParams containerParameters, enableMetrics bool, externalConfig *string) []corev1.Container {
+// generateContainerDef generates container definition for Redis
+func generateContainerDef(name string, containerParams containerParameters, enableMetrics bool, externalConfig *string, sidecars []redisv1beta1.Sidecar) []corev1.Container {
 	containerDefinition := []corev1.Container{
 		{
 			Name:            name,
 			Image:           containerParams.Image,
 			ImagePullPolicy: containerParams.ImagePullPolicy,
-			Env:             getEnvironmentVariables(containerParams.Role, containerParams.EnabledPassword, containerParams.SecretName, containerParams.SecretKey, containerParams.PersistenceEnabled, containerParams.RedisExporterEnv),
-			ReadinessProbe:  getProbeInfo(),
-			LivenessProbe:   getProbeInfo(),
-			VolumeMounts:    getVolumeMount(name, containerParams.PersistenceEnabled, externalConfig),
+			Env: getEnvironmentVariables(
+				containerParams.Role,
+				false,
+				containerParams.EnabledPassword,
+				containerParams.SecretName,
+				containerParams.SecretKey,
+				containerParams.PersistenceEnabled,
+				containerParams.RedisExporterEnv,
+				containerParams.TLSConfig,
+			),
+			ReadinessProbe: getProbeInfo(containerParams.ReadinessProbe),
+			LivenessProbe:  getProbeInfo(containerParams.LivenessProbe),
+			VolumeMounts:   getVolumeMount(name, containerParams.PersistenceEnabled, externalConfig, containerParams.TLSConfig),
 		},
 	}
+
 	if containerParams.Resources != nil {
 		containerDefinition[0].Resources = *containerParams.Resources
 	}
 	if enableMetrics {
 		containerDefinition = append(containerDefinition, enableRedisMonitoring(containerParams))
 	}
+	for _, sidecar := range sidecars {
+		container := corev1.Container{
+			Name:            sidecar.Name,
+			Image:           sidecar.Image,
+			ImagePullPolicy: sidecar.ImagePullPolicy,
+		}
+		if sidecar.Resources != nil {
+			container.Resources = *sidecar.Resources
+		}
+		if sidecar.EnvVars != nil {
+			container.Env = *sidecar.EnvVars
+		}
+		containerDefinition = append(containerDefinition, container)
+	}
 	return containerDefinition
+}
+
+func GenerateTLSEnvironmentVariables(tlsconfig *redisv1beta1.TLSConfig) []corev1.EnvVar {
+	var envVars []corev1.EnvVar
+	root := "/tls/"
+
+	// get and set Defaults
+	caCert := "ca.crt"
+	tlsCert := "tls.crt"
+	tlsCertKey := "tls.key"
+
+	if tlsconfig.CaKeyFile != "" {
+		caCert = tlsconfig.CaKeyFile
+	}
+	if tlsconfig.CertKeyFile != "" {
+		tlsCert = tlsconfig.CertKeyFile
+	}
+	if tlsconfig.KeyFile != "" {
+		tlsCertKey = tlsconfig.KeyFile
+	}
+
+	envVars = append(envVars, corev1.EnvVar{
+		Name:  "TLS_MODE",
+		Value: "true",
+	})
+	envVars = append(envVars, corev1.EnvVar{
+		Name:  "REDIS_TLS_CA_KEY",
+		Value: path.Join(root, caCert),
+	})
+	envVars = append(envVars, corev1.EnvVar{
+		Name:  "REDIS_TLS_CERT",
+		Value: path.Join(root, tlsCert),
+	})
+	envVars = append(envVars, corev1.EnvVar{
+		Name:  "REDIS_TLS_CERT_KEY",
+		Value: path.Join(root, tlsCertKey),
+	})
+	return envVars
 }
 
 // enableRedisMonitoring will add Redis Exporter as sidecar container
@@ -190,7 +296,17 @@ func enableRedisMonitoring(params containerParameters) corev1.Container {
 		Name:            redisExporterContainer,
 		Image:           params.RedisExporterImage,
 		ImagePullPolicy: params.RedisExporterImagePullPolicy,
-		Env:             getEnvironmentVariables(params.Role, params.EnabledPassword, params.SecretName, params.SecretKey, params.PersistenceEnabled, params.RedisExporterEnv),
+		Env: getEnvironmentVariables(
+			params.Role,
+			true,
+			params.EnabledPassword,
+			params.SecretName,
+			params.SecretKey,
+			params.PersistenceEnabled,
+			params.RedisExporterEnv,
+			params.TLSConfig,
+		),
+		VolumeMounts: getVolumeMount("", nil, nil, params.TLSConfig), // We need/want the tls-certs but we DON'T need the PVC (if one is available)
 	}
 	if params.RedisExporterResources != nil {
 		exporterDefinition.Resources = *params.RedisExporterResources
@@ -199,34 +315,43 @@ func enableRedisMonitoring(params containerParameters) corev1.Container {
 }
 
 // getVolumeMount gives information about persistence mount
-func getVolumeMount(name string, persistenceEnabled *bool, externalConfig *string) []corev1.VolumeMount {
-	var volumeMounts []corev1.VolumeMount
+func getVolumeMount(name string, persistenceEnabled *bool, externalConfig *string, tlsConfig *redisv1beta1.TLSConfig) []corev1.VolumeMount {
+	var VolumeMounts []corev1.VolumeMount
+
 	if persistenceEnabled != nil && *persistenceEnabled {
-		volumeMounts = []corev1.VolumeMount{
-			{
-				Name:      name,
-				MountPath: "/data",
-			},
-		}
+		VolumeMounts = append(VolumeMounts, corev1.VolumeMount{
+			Name:      name,
+			MountPath: "/data",
+		})
+	}
+
+	if tlsConfig != nil {
+		VolumeMounts = append(VolumeMounts, corev1.VolumeMount{
+			Name:      "tls-certs",
+			ReadOnly:  true,
+			MountPath: "/tls",
+		})
 	}
 
 	if externalConfig != nil {
-		volumeMounts = append(volumeMounts, corev1.VolumeMount{
+		VolumeMounts = append(VolumeMounts, corev1.VolumeMount{
 			Name:      "external-config",
 			MountPath: "/etc/redis/external.conf.d",
 		})
 	}
-	return volumeMounts
+
+	return VolumeMounts
 }
 
-// getProbeInfo generates probe information for Redis
-func getProbeInfo() *corev1.Probe {
+// getProbeInfo generate probe for Redis StatefulSet
+func getProbeInfo(probe *redisv1beta1.Probe) *corev1.Probe {
 	return &corev1.Probe{
-		InitialDelaySeconds: graceTime,
-		PeriodSeconds:       15,
-		FailureThreshold:    5,
-		TimeoutSeconds:      5,
-		Handler: corev1.Handler{
+		InitialDelaySeconds: probe.InitialDelaySeconds,
+		PeriodSeconds:       probe.PeriodSeconds,
+		FailureThreshold:    probe.FailureThreshold,
+		TimeoutSeconds:      probe.TimeoutSeconds,
+		SuccessThreshold:    probe.SuccessThreshold,
+		ProbeHandler: corev1.ProbeHandler{
 			Exec: &corev1.ExecAction{
 				Command: []string{
 					"bash",
@@ -238,12 +363,41 @@ func getProbeInfo() *corev1.Probe {
 }
 
 // getEnvironmentVariables returns all the required Environment Variables
-func getEnvironmentVariables(role string, enabledPassword *bool, secretName *string, secretKey *string, persistenceEnabled *bool, extraEnv *[]corev1.EnvVar) []corev1.EnvVar {
+func getEnvironmentVariables(role string, enabledMetric bool, enabledPassword *bool, secretName *string, secretKey *string, persistenceEnabled *bool, extraEnv *[]corev1.EnvVar, tlsConfig *redisv1beta1.TLSConfig) []corev1.EnvVar {
 	envVars := []corev1.EnvVar{
 		{Name: "SERVER_MODE", Value: role},
 		{Name: "SETUP_MODE", Value: role},
-		{Name: "REDIS_ADDR", Value: "redis://localhost:6379"},
 	}
+
+	redisHost := "redis://localhost:6379"
+	if tlsConfig != nil {
+		redisHost = "rediss://localhost:6379"
+		envVars = append(envVars, GenerateTLSEnvironmentVariables(tlsConfig)...)
+		if enabledMetric {
+			envVars = append(envVars, corev1.EnvVar{
+				Name:  "REDIS_EXPORTER_TLS_CLIENT_KEY_FILE",
+				Value: "/tls/tls.key",
+			})
+			envVars = append(envVars, corev1.EnvVar{
+				Name:  "REDIS_EXPORTER_TLS_CLIENT_CERT_FILE",
+				Value: "/tls/tls.crt",
+			})
+			envVars = append(envVars, corev1.EnvVar{
+				Name:  "REDIS_EXPORTER_TLS_CA_CERT_FILE",
+				Value: "/tls/ca.crt",
+			})
+			envVars = append(envVars, corev1.EnvVar{
+				Name:  "REDIS_EXPORTER_SKIP_TLS_VERIFICATION",
+				Value: "true",
+			})
+		}
+	}
+
+	envVars = append(envVars, corev1.EnvVar{
+		Name:  "REDIS_ADDR",
+		Value: redisHost,
+	})
+
 	if enabledPassword != nil && *enabledPassword {
 		envVars = append(envVars, corev1.EnvVar{
 			Name: "REDIS_PASSWORD",
@@ -270,9 +424,9 @@ func getEnvironmentVariables(role string, enabledPassword *bool, secretName *str
 	return envVars
 }
 
-// createStateFulSet is a method to create statefulset in Kubernetes
-func createStateFulSet(namespace string, stateful *appsv1.StatefulSet) error {
-	logger := stateFulSetLogger(namespace, stateful.Name)
+// createStatefulSet is a method to create statefulset in Kubernetes
+func createStatefulSet(namespace string, stateful *appsv1.StatefulSet) error {
+	logger := statefulSetLogger(namespace, stateful.Name)
 	_, err := generateK8sClient().AppsV1().StatefulSets(namespace).Create(context.TODO(), stateful, metav1.CreateOptions{})
 	if err != nil {
 		logger.Error(err, "Redis stateful creation failed")
@@ -282,32 +436,43 @@ func createStateFulSet(namespace string, stateful *appsv1.StatefulSet) error {
 	return nil
 }
 
-// updateStateFulSet is a method to update statefulset in Kubernetes
-func updateStateFulSet(namespace string, stateful *appsv1.StatefulSet) error {
-	logger := stateFulSetLogger(namespace, stateful.Name)
+// updateStatefulSet is a method to update statefulset in Kubernetes
+func updateStatefulSet(namespace string, stateful *appsv1.StatefulSet) error {
+	logger := statefulSetLogger(namespace, stateful.Name)
+	// logger.Info(fmt.Sprintf("Setting Statefulset to the following: %s", stateful))
 	_, err := generateK8sClient().AppsV1().StatefulSets(namespace).Update(context.TODO(), stateful, metav1.UpdateOptions{})
 	if err != nil {
 		logger.Error(err, "Redis stateful update failed")
 		return err
 	}
-	logger.Info("Redis stateful ")
+	logger.Info("Redis stateful successfully updated ")
 	return nil
 }
 
 // GetStateFulSet is a method to get statefulset in Kubernetes
-func GetStateFulSet(namespace string, stateful string) (*appsv1.StatefulSet, error) {
-	logger := stateFulSetLogger(namespace, stateful)
-	statefulInfo, err := generateK8sClient().AppsV1().StatefulSets(namespace).Get(context.TODO(), stateful, metav1.GetOptions{})
+func GetStatefulSet(namespace string, stateful string) (*appsv1.StatefulSet, error) {
+	logger := statefulSetLogger(namespace, stateful)
+	getOpts := metav1.GetOptions{
+		TypeMeta: generateMetaInformation("StatefulSet", "apps/v1"),
+	}
+	statefulInfo, err := generateK8sClient().AppsV1().StatefulSets(namespace).Get(context.TODO(), stateful, getOpts)
 	if err != nil {
 		logger.Info("Redis statefulset get action failed")
 		return nil, err
 	}
 	logger.Info("Redis statefulset get action was successful")
-	return statefulInfo, err
+	return statefulInfo, nil
 }
 
-// stateFulSetLogger will generate logging interface for Statfulsets
-func stateFulSetLogger(namespace string, name string) logr.Logger {
-	reqLogger := log.WithValues("Request.StateFulSet.Namespace", namespace, "Request.StateFulSet.Name", name)
+// statefulSetLogger will generate logging interface for Statfulsets
+func statefulSetLogger(namespace string, name string) logr.Logger {
+	reqLogger := log.WithValues("Request.StatefulSet.Namespace", namespace, "Request.StatefulSet.Name", name)
 	return reqLogger
+}
+
+func getSidecars(sidecars *[]redisv1beta1.Sidecar) []redisv1beta1.Sidecar {
+	if sidecars == nil {
+		return []redisv1beta1.Sidecar{}
+	}
+	return *sidecars
 }